requirements.txt
flask
redis
python-dotenv
pynetbox
netmiko
